Software/Diane Whisplay/requirements.txt
colorzero==2.0
gpiozero==2.0.1
mutagen==1.47.0
packaging==26.0
pillow==12.1.1
pygame==2.6.1
setuptools==82.0.0
spidev==3.8
wheel==0.46.3
